+++ test/fixtures/cookbooks/bind_test/recipes/spec_primary_zone.rb
@@ -4,12 +4,14 @@ end
 
 bind_config 'default'
 
-bind_primary_zone 'example.com'
-
 bind_primary_zone 'example.org' do
   options [
     'allow-transfer { none; }',
   ]
+end
+
+bind_primary_zone 'example.com' do
+  source_file 'custom-example.com'
 end
 
 bind_primary_zone 'example.net' do