+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of the bind cookbook.
 
 ## Unreleased
 
+- Add source file parameter to `bind_primary_zone`
+
 ## 3.0.2 - *2021-10-13*
 
 - Convert `node['platform_version']` to a float for correct comparison